Make Generic HTTP Request more robust and flexible
Update Generic HTTP Request to pass request and response classes to constructor
This solves the issue requireing a Wrapper class and a ParameterizedTypeThing
when calling the requestTemplate execute function.

It also means that whatever Model you pass to the contructor (User.class,
Success.class, ect) is exactly what will be parsed to by Jackson2

This enables us to recieve a response of JSON that describes the model directly
instead of requiring the response contain a model wrapped in another JSON key.